Removed measurement form & added individual forms
